+++ BasicRenderer/src/mesh.cpp
@@ -12,7 +12,7 @@
 #include "MeshManager.h"
 #include "Skeleton.h"
 
-std::atomic<int> Mesh::globalMeshCount = 0;
+std::atomic<uint64_t> Mesh::globalMeshCount = 0;
 
 Mesh::Mesh(std::unique_ptr<std::vector<std::byte>> vertices, unsigned int vertexSize, std::unique_ptr<std::vector<std::byte>> skinningVertices, unsigned int skinningVertexSize, const std::vector<UINT32>& indices, const std::shared_ptr<Material> material, unsigned int flags) {
     m_vertices = std::move(vertices);
@@ -136,7 +136,7 @@ int Mesh::GetNextGlobalIndex() {
     return globalMeshCount.fetch_add(1, std::memory_order_relaxed);
 }
 
-int Mesh::GetGlobalID() const {
+uint64_t Mesh::GetGlobalID() const {
 	return m_globalMeshID;
 }
